Amend default pod image to be non-existent image
If the image name is not changed to something different
from what is defined in pod.yaml then the apply
`kubectl set image` command does not return a changed
object to pass to the apply command.

To support deploy of branches and master we
must, therefore, provide a placeholder image tag
that MUST be 'set' as part of deploy.